Reformat code with OOP
Reformatted code to extract out Ui, Storage, Parser and TaskList into each of their own classes
