+++ src/main/java/GopsException.java
@@ -1,0 +1,6 @@
+public class GopsException extends Exception {
+    public GopsException() {
+        super();
+    }
+
+}
