--- src/main/java/gopsException.java
@@ -1,6 +1,0 @@
-public class gopsException extends Exception {
-    public gopsException() {
-        super();
-    }
-
-}
